// front-web/src/pages/records/types.ts
export type RecordsResponse = {
  content: RecordItem[];
  totalPages: number;
};

export type RecordItem = {
  id: number;
  moment: string;
  email: string;
  name: string;
  age: number;
  gameTitle: string;
  gamePlatform: Platform;
  genreName: string;
};

export type Platform = 'XBOX' | 'NINTENDO' | 'PLAYSTATION';